--- a/data.sql
+++ b/data.sql
@@ -130,3 +130,10 @@ VALUES(7, 2, DATE '2020-02-27');
 VALUES(7, 2, DATE '2020-08-03');
 VALUES(8, 4, DATE '2020-05-24');
 VALUES(8, 1, DATE '2021-01-11');
+
+/*Database performance audit*/
+
+
+INSERT INTO visits (animals_id, vets_id, date_of_visit) SELECT * FROM (SELECT id FROM animals) animals_ids, (SELECT id FROM vets) vets_ids, generate_series('1980-01-01'::timestamp, '2021-01-01', '4 hours') visit_timestamp;
+
+insert into owners (full_name, email) select 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com';