--- a/queries.sql
+++ b/queries.sql
@@ -118,3 +118,6 @@ FROM animals JOIN species ON animals.species_id = species.id JOIN visits ON anim
 WHERE vets.name = 'Vet Maisy Smith'
 GROUP BY species.name
 LIMIT 1;
+
+/* Database performance audit*/
+explain analyze SELECT COUNT(*) FROM visits where animals_id = 4;