--- a/queries.sql
+++ b/queries.sql
@@ -121,3 +121,4 @@ LIMIT 1;
 
 /* Database performance audit*/
 explain analyze SELECT COUNT(*) FROM visits where animals_id = 4;
+explain analyze SELECT * FROM visits where vets_id = 2;