--- a/queries.sql
+++ b/queries.sql
@@ -120,6 +120,6 @@ GROUP BY species.name
 LIMIT 1;
 
 /* Database performance audit*/
-SELECT COUNT(*) FROM visits where animal_id = 4;
-SELECT * FROM visits where vet_id = 2;
+SELECT COUNT(*) FROM visits where animals_id = 4;
+SELECT * FROM visits where vets_id = 2;
 SELECT * FROM owners where email = 'owner_18327@mail.com';