--- a/schema.sql
+++ b/schema.sql
@@ -57,3 +57,8 @@ CREATE TABLE visits (
 	vets_id BIGINT REFERENCES vets (id),
 	date_of_visit DATE NOT NULL
 )
+
+
+/* Database performance audit*/
+
+ALTER TABLE owners ADD COLUMN email VARCHAR(120);