--- a/schema.sql
+++ b/schema.sql
@@ -62,3 +62,5 @@ CREATE TABLE visits (
 /* Database performance audit*/
 
 ALTER TABLE owners ADD COLUMN email VARCHAR(120);
+
+INSERT INTO visits (animals_id, vets_id, date_of_visit) SELECT * FROM (SELECT id FROM animals) animals_ids, (SELECT id FROM vets) vets_ids, generate_series('1980-01-01'::timestamp, '2021-01-01', '4 hours') visit_timestamp;