--- a/schema.sql
+++ b/schema.sql
@@ -64,6 +64,6 @@ CREATE TABLE visits (
 ALTER TABLE owners ADD COLUMN email VARCHAR(120);
 
 -- Indexes
-CREATE INDEX animals_id_asc ON visits(animal_id ASC);
+CREATE INDEX animals_id_asc ON visits(animals_id ASC);
 CREATE INDEX vets_id_asc ON visits(vets_id ASC);
 CREATE INDEX email_asc ON owners(email ASC);