… images/owners-table-analyzer more than 1000 ms.PNG
… images/owners-table-analyzer-less-than-1000 ms.PNG
… images/visits-table-analyzer 1000 ms.PNG
… images/visits-table-analyzer-1 1000 ms.PNG
… images/visits-table-analyzer-1-less-than-1000 ms.PNG
… images/visits-table-analyzer-2-less-than-1000 ms.PNG