# Database performance audit
Add indexes to improve queries performance.

| Before         | After     |
|--------------|-----------|
|![visits-table-analyzer 1000 ms](https://user-images.githubusercontent.com/71156605/150872881-66a20b85-9ba5-464f-8fc6-9d48e8e8d9f3.PNG)      | ![visits-table-analyzer-1-less-than-1000 ms](https://user-images.githubusercontent.com/71156605/150872887-1f05563d-0ae8-435c-a71c-e52bc44bd0b8.PNG)    | 
|![visits-table-analyzer-1 1000 ms](https://user-images.githubusercontent.com/71156605/150872884-8b520b1e-be9f-4988-9c0d-ec42655b61f1.PNG)|![visits-table-analyzer-2-less-than-1000 ms](https://user-images.githubusercontent.com/71156605/150872888-77d31a2a-fb31-4f01-980f-38ee3eb70d70.PNG)|
|![owners-table-analyzer more than 1000 ms](https://user-images.githubusercontent.com/71156605/150872876-d12090f6-2b14-4af0-a6b2-6433f619d58e.PNG) |![owners-table-analyzer-less-than-1000 ms](https://user-images.githubusercontent.com/71156605/150872871-48c10820-40fe-4633-92a5-cdc57f5c0439.PNG) |